+++ cloudapi-20160714/alibabacloud_cloudapi20160714/models.py
@@ -19882,6 +19882,7 @@ class DescribeAppResponseBody(TeaModel):
         app_name: str = None,
         created_time: str = None,
         description: str = None,
+        disabled: bool = None,
         extend: str = None,
         modified_time: str = None,
         request_id: str = None,
@@ -19894,6 +19895,7 @@ class DescribeAppResponseBody(TeaModel):
         self.created_time = created_time
         # The description of the app.
         self.description = description
+        self.disabled = disabled
         # 扩展信息
         self.extend = extend
         # The time when the app was modified.
@@ -19918,6 +19920,8 @@ class DescribeAppResponseBody(TeaModel):
             result['CreatedTime'] = self.created_time
         if self.description is not None:
             result['Description'] = self.description
+        if self.disabled is not None:
+            result['Disabled'] = self.disabled
         if self.extend is not None:
             result['Extend'] = self.extend
         if self.modified_time is not None:
@@ -19936,6 +19940,8 @@ class DescribeAppResponseBody(TeaModel):
             self.created_time = m.get('CreatedTime')
         if m.get('Description') is not None:
             self.description = m.get('Description')
+        if m.get('Disabled') is not None:
+            self.disabled = m.get('Disabled')
         if m.get('Extend') is not None:
             self.extend = m.get('Extend')
         if m.get('ModifiedTime') is not None:
@@ -20212,6 +20218,7 @@ class DescribeAppAttributesResponseBodyAppsAppAttribute(TeaModel):
         app_name: str = None,
         created_time: str = None,
         description: str = None,
+        disabled: bool = None,
         extend: str = None,
         modified_time: str = None,
         tags: DescribeAppAttributesResponseBodyAppsAppAttributeTags = None,
@@ -20224,6 +20231,7 @@ class DescribeAppAttributesResponseBodyAppsAppAttribute(TeaModel):
         self.created_time = created_time
         # The description of the app.
         self.description = description
+        self.disabled = disabled
         # 扩展信息
         self.extend = extend
         # The modification time (UTC) of the app.
@@ -20249,6 +20257,8 @@ class DescribeAppAttributesResponseBodyAppsAppAttribute(TeaModel):
             result['CreatedTime'] = self.created_time
         if self.description is not None:
             result['Description'] = self.description
+        if self.disabled is not None:
+            result['Disabled'] = self.disabled
         if self.extend is not None:
             result['Extend'] = self.extend
         if self.modified_time is not None:
@@ -20267,6 +20277,8 @@ class DescribeAppAttributesResponseBodyAppsAppAttribute(TeaModel):
             self.created_time = m.get('CreatedTime')
         if m.get('Description') is not None:
             self.description = m.get('Description')
+        if m.get('Disabled') is not None:
+            self.disabled = m.get('Disabled')
         if m.get('Extend') is not None:
             self.extend = m.get('Extend')
         if m.get('ModifiedTime') is not None:
@@ -43155,6 +43167,7 @@ class ModifyAppRequest(TeaModel):
         app_id: int = None,
         app_name: str = None,
         description: str = None,
+        disabled: bool = None,
         extend: str = None,
         security_token: str = None,
         tag: List[ModifyAppRequestTag] = None,
@@ -43171,6 +43184,7 @@ class ModifyAppRequest(TeaModel):
         # 
         # This parameter is required only when you want to modify the value.
         self.description = description
+        self.disabled = disabled
         # 扩展信息
         self.extend = extend
         self.security_token = security_token
@@ -43195,6 +43209,8 @@ class ModifyAppRequest(TeaModel):
             result['AppName'] = self.app_name
         if self.description is not None:
             result['Description'] = self.description
+        if self.disabled is not None:
+            result['Disabled'] = self.disabled
         if self.extend is not None:
             result['Extend'] = self.extend
         if self.security_token is not None:
@@ -43213,6 +43229,8 @@ class ModifyAppRequest(TeaModel):
             self.app_name = m.get('AppName')
         if m.get('Description') is not None:
             self.description = m.get('Description')
+        if m.get('Disabled') is not None:
+            self.disabled = m.get('Disabled')
         if m.get('Extend') is not None:
             self.extend = m.get('Extend')
         if m.get('SecurityToken') is not None: